+++ lib/src/domain/enums/enums.dart
@@ -1,3 +1,3 @@
-export 'validation_mode.dart';
+export 'validation_trigger.dart';
 export 'error_reset_mode.dart';
 export 'field_status.dart';