+++ lib/src/domain/enums/validation_mode.dart
@@ -1,5 +1,0 @@
-enum ValidationMode {
-  valueChange,
-  focusChange,
-  noValidate,
-}